Issue/22 (#23)
* Feat: BuildingException 및 ErrorCode 작성

* Feat: jacoco 설치 및 코드 커버리지 확인 & 필요없는 테스트 삭제

* Feat: buildingName 추가 및 buildingSummary 연관관계 설정
- Building에 buildingName 추가
- BuildingSummary와 Building 간 단방향 일대일 연관관계
- BuildingSummary 컬럼 변경

* Feat: RoomResponseDto 내의 Id 값들 전부 Long으로

* Add: Projection 사용해서 findBy를 이용해서 buildingId와 Coordinate만 검색해온 후 controller에서 return

* Add: BuildingRestController 구현

* Fix: avgScore 타입 변화

* Add: Serializer 작성

* Add: 기존 Serializer 제거 및 rebase 꼬인 거 해결

* Add: create BuildingToReviewCategoryRepository

* Add: apply Serializer to BuildingRestController

* Add: change code block of getBuildingList in that filling the buildingScoreAvg and bestCategory with using buildingToReviewCategory

* Fix: Controller와 Service에 Pageable 적용

* Fix: create ApiErrorResult for exceptionHandling

* Fix: make CustomException and make BuildingException extends it

* Fix: change All method to use ApiErrorResult as return body

* Add: make ExceptionController extend ResponseEntityExceptionHandler that handle almost exceptions in mvc automatically.
Therefore, we can deal with CustomException in same way with using handleExceptionInternal

+++ src/main/java/com/project/Project/controller/building/dto/BuildingResponseDto.java
@@ -22,7 +22,7 @@ public class BuildingResponseDto {
         private String name;
         private String address;
         private Long reviewCnt;
-        private BigDecimal scoreAvg;
+        private Double scoreAvg;
         private ReviewCategoryEnum bestCategory;
         private boolean isDirectDeal;
     }
@@ -50,7 +50,7 @@ public class BuildingResponseDto {
         private CoordinateDto coordinate;
         private Boolean isDirectDeal;
         private List<RoomResponseDto.RoomListResponse> rooms;
-        private Map<ReviewCategoryEnum, BigDecimal> buildingSummaries;
+        private Map<ReviewCategoryEnum, Double> buildingSummaries;
     }
 }
